requirements.txt
Flask==2.2.5
python-docx==0.8.11
WeasyPrint==57.1
Jinja2==3.1.2
